+++ src/services/user/follow.ts
@@ -4,7 +4,7 @@ import { UserInfoDto } from '../../interfaces/user/userInfo';
 import { FollowListDto } from '../../interfaces/user/userInfo';
 import { SingleNotificationResponse } from '../../interfaces/response';
 import { redis } from '../../loaders/redis';
-import { LimitRequestDto } from '../../interfaces/LimitRequestDto';
+import { LimitRequestDto } from '../../interfaces/limitRequestDto';
 import { CacheKeys } from '../../constants/cacheKeys';
 
 export class FollowService {
